Move geojson writer/reader to individual modules
